more refactoring for runs
