--- domino_cost/__init__.py
@@ -1,3 +1,0 @@
-__all__ = ["domino_cost", "constants"]
-
-from domino_cost import constants, domino_cost
