+++ domino_cost/requests_helpers.py
@@ -1,5 +1,3 @@
-from __future__ import annotations
-
 from typing import List
 
 import requests
